XML service + controller + Unit test
Bugs verwijderd in originele XMlconverter, controller gemaakt zodat 2 XML bestanden kunnen worden geupload zou via API call mogelijk zijn nu. Unit test in test environment gemaakt voor XML
