+++ ProjectD/Models/Enums.cs
@@ -1,37 +1,40 @@
-public enum InventoryTransactionType
+namespace ProjectD.Models
 {
-    Inbound,
-    Outbound,
-    Adjustment
-}
+    public enum InventoryTransactionType
+    {
+        Inbound,
+        Outbound,
+        Adjustment
+    }
 
-public enum OrderStatus
-{
-    Pending,
-    Processing,
-    Shipped,
-    Delivered,
-    Cancelled
-}
+    public enum OrderStatus
+    {
+        Pending,
+        Processing,
+        Shipped,
+        Delivered,
+        Cancelled
+    }
 
-public enum VehicleType
-{
-    FlatbedTrailer,
-    LowbedTrailer,
-    Kipper
-}
+    public enum VehicleType
+    {
+        FlatbedTrailer,
+        LowbedTrailer,
+        Kipper
+    }
 
-public enum VehicleStatus
-{
-    Available,
-    InUse,
-    Maintenance
-}
+    public enum VehicleStatus
+    {
+        Available,
+        InUse,
+        Maintenance
+    }
 
-public enum ShipmentStatus
-{
-    Preparing,
-    OutForDelivery,
-    Delivered
-}
+    public enum ShipmentStatus
+    {
+        Preparing,
+        OutForDelivery,
+        Delivered
+    }
 
+}
