+++ ProjectD/Models/Order.cs
@@ -1,35 +1,37 @@
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
-using AzureSqlConnectionDemo.Models;
+using ProjectD.Models;
 
-
-public class Order
+namespace ProjectD.Models
 {
-    [Key]
-    public int Id { get; set; }
+    public class Order
+    {
+        [Key]
+        public int Id { get; set; }
 
-    [Required]
-    public int CustomerId { get; set; }
+        [Required]
+        public int CustomerId { get; set; }
 
-    [ForeignKey("CustomerId")]
-    public Customer Customer { get; set; } = default!;
+        [ForeignKey("CustomerId")]
+        public Customer Customer { get; set; } = default!;
 
-    public List<OrderLine> ProductLines { get; set; } = new();
+        public List<OrderLine> ProductLines { get; set; } = new();
 
-    [NotMapped]
-    public int TotalWeight => ProductLines.Sum(pl => pl.Product.WeightKg * pl.Quantity);
+        [NotMapped]
+        public int TotalWeight => ProductLines.Sum(pl => pl.Product.WeightKg * pl.Quantity);
 
-    public OrderStatus Status { get; set; }
+        public OrderStatus Status { get; set; }
 
-    public DateTime OrderDate { get; set; }
+        public DateTime OrderDate { get; set; }
 
-    public string DeliveryAddress { get; set; } = string.Empty;
+        public string DeliveryAddress { get; set; } = string.Empty;
 
-    public DateTime ExpectedDeliveryDate { get; set; }
+        public DateTime ExpectedDeliveryDate { get; set; }
 
-    public DateTime? ActualDeliveryDate { get; set; }
+        public DateTime? ActualDeliveryDate { get; set; }
 
-    public bool IsDeleted { get; set; }
-    public ICollection<ShipmentOrder> ShipmentOrders { get; set; } = new List<ShipmentOrder>();
+        public bool IsDeleted { get; set; }
+        public ICollection<ShipmentOrder> ShipmentOrders { get; set; } = new List<ShipmentOrder>();
 
+    }
 }
