+++ ProjectD/Models/OrderLine.cs
@@ -1,24 +1,26 @@
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
-
-public class OrderLine
+namespace ProjectD.Models
 {
-    [Key]
-    public int Id { get; set; }
+    public class OrderLine
+    {
+        [Key]
+        public int Id { get; set; }
 
-    public int OrderId { get; set; }
+        public int OrderId { get; set; }
 
-    [ForeignKey("OrderId")]
-    public Order Order { get; set; } = default!;
+        [ForeignKey("OrderId")]
+        public Order Order { get; set; } = default!;
 
-    public int ProductId { get; set; }
+        public int ProductId { get; set; }
 
-    [ForeignKey("ProductId")]
-    public Product Product { get; set; } = default!;
+        [ForeignKey("ProductId")]
+        public Product Product { get; set; } = default!;
 
-    public int Quantity { get; set; }
+        public int Quantity { get; set; }
 
-    [NotMapped]
-    public double LineTotal => Product.Price * Quantity;
-    public bool IsDeleted { get; set; }
+        [NotMapped]
+        public double LineTotal => Product.Price * Quantity;
+        public bool IsDeleted { get; set; }
+    }
 }
