+++ ProjectD/Models/Product.cs
@@ -1,26 +1,28 @@
 using System.ComponentModel.DataAnnotations;
-
-public class Product
+namespace ProjectD.Models
 {
-    [Key]
-    public int Id { get; set; }
+    public class Product
+    {
+        [Key]
+        public int Id { get; set; }
 
-    [Required]
-    public string ProductName { get; set; } = string.Empty;
+        [Required]
+        public string ProductName { get; set; } = string.Empty;
 
-    public string SKU { get; set; } = string.Empty;
+        public string SKU { get; set; } = string.Empty;
 
-    public int WeightKg { get; set; }
+        public int WeightKg { get; set; }
 
-    public string Material { get; set; } = string.Empty;
+        public string Material { get; set; } = string.Empty;
 
-    public int BatchNumber { get; set; }
+        public int BatchNumber { get; set; }
 
-    [Range(0, double.MaxValue)]
-    public double Price { get; set; }
+        [Range(0, double.MaxValue)]
+        public double Price { get; set; }
 
-    public string Category { get; set; } = string.Empty;
+        public string Category { get; set; } = string.Empty;
 
-    public DateTime? ExpirationDate { get; set; }
-    public bool IsDeleted { get; set; }
+        public DateTime? ExpirationDate { get; set; }
+        public bool IsDeleted { get; set; }
+    }
 }
