+++ ProjectD/Models/SeedData.cs
@@ -1,8 +1,9 @@
 using Microsoft.EntityFrameworkCore;
 using System;
 using System.Linq;
+using ProjectD.Models;
 
-namespace AzureSqlConnectionDemo.Models
+namespace ProjectD.Models
 {
     public static class SeedData
     {
