+++ ProjectD/Models/Vehicle.cs
@@ -1,16 +1,19 @@
 using System.ComponentModel.DataAnnotations;
 
-public class Vehicle
+namespace ProjectD.Models
 {
-    [Key]
-    public int Id { get; set; }
+    public class Vehicle
+    {
+        [Key]
+        public int Id { get; set; }
 
-    public string LicensePlate { get; set; } = string.Empty;
+        public string LicensePlate { get; set; } = string.Empty;
 
-    public int CapacityKg { get; set; }
+        public int CapacityKg { get; set; }
 
-    public VehicleType Type { get; set; }
+        public VehicleType Type { get; set; }
 
-    public VehicleStatus Status { get; set; }
-    public bool IsDeleted { get; set; }
+        public VehicleStatus Status { get; set; }
+        public bool IsDeleted { get; set; }
+    }
 }
